+++ src/RobotStudio.Desktop/Showcases/MechanicalShowcaseCatalog.cs
@@ -8,7 +8,8 @@ internal static class MechanicalShowcaseCatalog
             ["cartesian-intro-mechanical"] = CartesianMechanicalShowcaseDefinition.CreatePresentation,
             ["xy-plotter-mechanical"] = XYPlotterMechanicalShowcaseDefinition.CreatePresentation,
             ["differential-drive-mechanical"] = DifferentialDriveMechanicalShowcaseDefinition.CreatePresentation,
-            ["scara-mechanical"] = ScaraMechanicalShowcaseDefinition.CreatePresentation
+            ["scara-mechanical"] = ScaraMechanicalShowcaseDefinition.CreatePresentation,
+            ["simple-arm-mechanical"] = SimpleArmMechanicalShowcaseDefinition.CreatePresentation
         };
 
     public static IReadOnlyList<string> ModelIds { get; } = Factories.Keys.ToArray();